// src/typings/user.type.ts
export type UserAPIResponse = {
  id: string;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
};

export type UserNode = {
  id: UserAPIResponse["id"];
  name: UserAPIResponse["name"];
  email: UserAPIResponse["email"];
  city: UserAPIResponse["address"]["city"];
  companyName: UserAPIResponse["company"]["name"];
};